// generated/model/vedleggOpplastingResponseStatus.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type VedleggOpplastingResponseStatus =
    (typeof VedleggOpplastingResponseStatus)[keyof typeof VedleggOpplastingResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VedleggOpplastingResponseStatus = {
    OK: "OK",
    COULD_NOT_LOAD_DOCUMENT: "COULD_NOT_LOAD_DOCUMENT",
    PDF_IS_ENCRYPTED: "PDF_IS_ENCRYPTED",
    ILLEGAL_FILE_TYPE: "ILLEGAL_FILE_TYPE",
    ILLEGAL_FILENAME: "ILLEGAL_FILENAME",
    FILE_TOO_LARGE: "FILE_TOO_LARGE",
} as const;